requirements.txt
streamlit==1.31.1
langchain==0.1.4
langchain-community==0.0.16
python-dotenv==1.0.1
google-api-python-client==2.116.0
beautifulsoup4==4.12.3
selenium==4.17.2
chromadb==0.4.22
openai==1.12.0
faiss-cpu==1.7.4
tiktoken==0.5.2
